Update README.md
Updated the readme.md similarly to how a digital notebook entry is done, I gave a description of the code, saying what my code does and presents to the user on the website. Additionally, at the bottom I added a description of the code I used, including the Inline Styling.

Finally, at the bottom, I included a preview of the about-me website at the bottom of the readme.md.